+++ server.js
@@ -5,7 +5,7 @@ var fs = require('fs');
 var zlib = require('zlib');
 var path = require('path');
 var mime = require('mime');
-var webvtt = require('./scripts/webvtt');
+var webvtt = require('./modules/webvtt');
 var client = require('./modules/redis');
 var mailer = require('./modules/mailer');
 var spawn = require('child_process').spawn;
